requirements.txt
opencv-python>=4.5.0
Pillow>=10.0.0
PySide6>=6.5.0
imagehash>=4.3.0
numpy>=1.21.0

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0